AP-1724: Force to set fixed height in icon button when size changed.

See merge request desktop/desktop!3049

+++ src/MEGASync/control/ImageDownloader.cpp
@@ -42,7 +42,7 @@ void ImageDownloader::downloadImage(const QString& imageUrl,
     }
 
     QNetworkRequest request(url);
-    request.setTransferTimeout(mTimeout);
+    request.setTransferTimeout(static_cast<int>(mTimeout));
     request.setAttribute(QNetworkRequest::CacheLoadControlAttribute, QNetworkRequest::AlwaysNetwork);
 
     QNetworkReply* reply = mManager->get(request);
